put simd specific operator validation in separate file
